move temporary settings to shared
