+++ client/search/.eslintignore
@@ -1,1 +1,2 @@
 out/
+src/graphql-operations.ts
