+++ client/shared/src/settings/temporary/searchSidebar.ts
@@ -1,0 +1,16 @@
+export enum SectionID {
+    SEARCH_REFERENCE = 'reference',
+    SEARCH_TYPES = 'types',
+    DYNAMIC_FILTERS = 'filters',
+    REPOSITORIES = 'repositories',
+    SEARCH_SNIPPETS = 'snippets',
+    QUICK_LINKS = 'quicklinks',
+    REVISIONS = 'revisions',
+}
+
+export enum NoResultsSectionID {
+    SEARCH_BAR = 'search-bar',
+    LITERAL_SEARCH = 'literal-search',
+    COMMON_PROBLEMS = 'common-problems',
+    VIDEOS = 'videos',
+}
