+++ client/web/src/enterprise/insights/InsightsRouter.tsx
@@ -5,6 +5,7 @@ import { RouteComponentProps, Switch, Route, useRouteMatch } from 'react-router'
 import { Redirect } from 'react-router-dom'
 
 import { PlatformContextProps } from '@sourcegraph/shared/src/platform/context'
+import { Settings } from '@sourcegraph/shared/src/schema/settings.schema'
 import { SettingsCascadeProps } from '@sourcegraph/shared/src/settings/settings'
 import { TelemetryProps } from '@sourcegraph/shared/src/telemetry/telemetryService'
 
@@ -12,7 +13,6 @@ import { AuthenticatedUser } from '../../auth'
 import { withAuthenticatedUser } from '../../auth/withAuthenticatedUser'
 import { HeroPage } from '../../components/HeroPage'
 import { CodeInsightsContextProps } from '../../insights/types'
-import { Settings } from '@sourcegraph/shared/src/schema/settings.schema'
 import { lazyComponent } from '../../util/lazyComponent'
 
 import { CodeInsightsBackendContext } from './core/backend/code-insights-backend-context'
