+++ client/web/src/enterprise/insights/core/backend/setting-based-api/utils/delete-helpers.ts
@@ -1,9 +1,9 @@
 import { get } from 'lodash'
 
 import { isErrorLike } from '@sourcegraph/common'
+import { Settings } from '@sourcegraph/shared/src/schema/settings.schema'
 import { SettingsCascadeProps } from '@sourcegraph/shared/src/settings/settings'
 
-import { Settings } from '@sourcegraph/shared/src/schema/settings.schema'
 import {
     RemoveInsight,
     RemoveInsightFromDashboard,
