+++ client/web/src/enterprise/insights/core/backend/setting-based-api/utils/find-insight-by-id.ts
@@ -1,7 +1,7 @@
 import { isErrorLike } from '@sourcegraph/common'
+import { Settings } from '@sourcegraph/shared/src/schema/settings.schema'
 import { ConfiguredSubjectOrError, SettingsCascadeOrError } from '@sourcegraph/shared/src/settings/settings'
 
-import { Settings } from '@sourcegraph/shared/src/schema/settings.schema'
 import {
     Insight,
     InsightExecutionType,
