+++ client/web/src/enterprise/insights/core/backend/setting-based-api/utils/get-insights-dashboards/get-insights-dashboards.ts
@@ -1,7 +1,7 @@
 import { ErrorLike, isErrorLike } from '@sourcegraph/common'
+import { Settings } from '@sourcegraph/shared/src/schema/settings.schema'
 import { ConfiguredSubjectOrError } from '@sourcegraph/shared/src/settings/settings'
 
-import { Settings } from '@sourcegraph/shared/src/schema/settings.schema'
 import { InsightDashboard } from '../../../../types'
 import { ALL_INSIGHTS_DASHBOARD } from '../../../../types/dashboard/virtual-dashboard'
 import { isSubjectInsightSupported } from '../../../../types/subjects'
