+++ client/web/src/enterprise/insights/core/backend/setting-based-api/utils/get-updated-subject-settings.ts
@@ -1,7 +1,7 @@
 import { isErrorLike } from '@sourcegraph/common'
+import { Settings } from '@sourcegraph/shared/src/schema/settings.schema'
 import { ConfiguredSubjectOrError, SettingsCascadeProps } from '@sourcegraph/shared/src/settings/settings'
 
-import { Settings } from '@sourcegraph/shared/src/schema/settings.schema'
 import {
     AddInsight,
     AddInsightToDashboard,
