+++ client/web/src/enterprise/insights/core/settings-action/dashboards.ts
@@ -1,9 +1,9 @@
 import { camelCase } from 'lodash'
 
 import { isErrorLike } from '@sourcegraph/common'
+import { InsightDashboard, Settings } from '@sourcegraph/shared/src/schema/settings.schema'
 import { modify, parseJSONCOrError } from '@sourcegraph/shared/src/util/jsonc'
 
-import { InsightDashboard, Settings } from '@sourcegraph/shared/src/schema/settings.schema'
 import { INSIGHTS_DASHBOARDS_SETTINGS_KEY } from '../types'
 
 /**
