+++ client/web/src/enterprise/insights/mocks/settings-cascade.ts
@@ -1,6 +1,5 @@
-import { SettingsCascade, SettingsSubject } from '@sourcegraph/shared/src/settings/settings'
-
 import { Settings } from '@sourcegraph/shared/src/schema/settings.schema'
+import { SettingsCascade, SettingsSubject } from '@sourcegraph/shared/src/settings/settings'
 
 export const SETTINGS_CASCADE_MOCK: SettingsCascade<Settings> = {
     subjects: [
