+++ client/web/src/insights/sections/types.ts
@@ -1,9 +1,8 @@
 import { ExtensionsControllerProps } from '@sourcegraph/shared/src/extensions/controller'
 import { PlatformContextProps } from '@sourcegraph/shared/src/platform/context'
+import { Settings } from '@sourcegraph/shared/src/schema/settings.schema'
 import { SettingsCascadeProps } from '@sourcegraph/shared/src/settings/settings'
 import { TelemetryProps } from '@sourcegraph/shared/src/telemetry/telemetryService'
-
-import { Settings } from '@sourcegraph/shared/src/schema/settings.schema'
 
 /**
  * Common props for extend point component for rendering extension views section.
