+++ client/web/src/insights/utils/is-code-insights-enabled.ts
@@ -1,7 +1,6 @@
 import { isErrorLike } from '@sourcegraph/common'
-import { SettingsCascadeOrError } from '@sourcegraph/shared/src/settings/settings'
-
 import { SettingsExperimentalFeatures } from '@sourcegraph/shared/src/schema/settings.schema'
+import { SettingsCascadeOrError } from '@sourcegraph/shared/src/settings/settings'
 
 /**
  * Code insights display location setting to check setting for particular view
