+++ client/web/src/search/input/SearchOnboardingTour.ts
@@ -11,8 +11,8 @@ import Tour from 'shepherd.js/src/types/tour'
 import { ALL_LANGUAGES } from '@sourcegraph/shared/src/search/query/languageFilter'
 import { scanSearchQuery } from '@sourcegraph/shared/src/search/query/scanner'
 import { Token } from '@sourcegraph/shared/src/search/query/token'
+import { useTemporarySetting } from '@sourcegraph/shared/src/settings/temporary/useTemporarySetting'
 
-import { useTemporarySetting } from '../../settings/temporary/useTemporarySetting'
 import { eventLogger } from '../../tracking/eventLogger'
 import { isMacPlatform } from '../../util'
 import { QueryState } from '../helpers'
