+++ client/web/src/search/results/sidebar/Revisions.tsx
@@ -7,6 +7,7 @@ import { QueryUpdate } from '@sourcegraph/search'
 import { SyntaxHighlightedSearchQuery } from '@sourcegraph/search-ui'
 import { GitRefType } from '@sourcegraph/shared/src/schema'
 import { FilterType } from '@sourcegraph/shared/src/search/query/filters'
+import { useTemporarySetting } from '@sourcegraph/shared/src/settings/temporary/useTemporarySetting'
 import { Button, LoadingSpinner } from '@sourcegraph/wildcard'
 
 import { useConnection } from '../../../components/FilteredConnection/hooks/useConnection'
@@ -15,7 +16,6 @@ import {
     SearchSidebarGitRefsVariables,
     SearchSidebarGitRefFields,
 } from '../../../graphql-operations'
-import { useTemporarySetting } from '../../../settings/temporary/useTemporarySetting'
 
 import { FilterLink } from './FilterLink'
 import styles from './SearchSidebarSection.module.scss'
